move script to command as skywire-cli log collecting
